+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+- `ListItemLayout` is now reusable in other lists, by specifying a layout resource to inflate.
+
 ## 0.1.0 - 2015-02-26
 ### Added
 - Initial release (unstable)